+++ python/lsst/ts/criopy/m1m3/ApplicationControlWidget.py
@@ -241,7 +241,12 @@ class M1M3CSCControl(CSCControlWidget):
     TEXT_ABORT_RAISE = "&Abort M1M3 Raise"
     TEXT_LOWER = "&Lower M1M3"
     TEXT_ENTER_ENGINEERING = "&Enter Engineering"
-    TEXT_EXIT_ENGINEERING = "&Exit Engineering"
+    TEXT_EXIT_ENGINEERING = "E&xit Engineering"
+    TEXT_PANIC = "&Panic!"
+    TEXT_PAUSE_RAISING = "Pause raising"
+    TEXT_RESUME_RAISING = "Resume raising"
+    TEXT_PAUSE_LOWERING = "Pause lowering"
+    TEXT_RESUME_LOWERING = "Resume lowering"
 
     def __init__(self, m1m3: MetaSAL):
         super().__init__(
@@ -250,24 +255,41 @@ class M1M3CSCControl(CSCControlWidget):
                 self.TEXT_START,
                 self.TEXT_ENABLE,
                 self.TEXT_RAISE,
+                self.TEXT_PAUSE_RAISING,
                 self.TEXT_ENTER_ENGINEERING,
                 self.TEXT_STANDBY,
             ],
             {
                 self.TEXT_RAISE: "raiseM1M3",
+                self.TEXT_PAUSE_RAISING: "pauseM1M3RaisingLowering",
+                self.TEXT_RESUME_RAISING: "resumeM1M3RaisingLowering",
                 self.TEXT_ABORT_RAISE: "abortRaiseM1M3",
                 self.TEXT_LOWER: "lowerM1M3",
+                self.TEXT_PAUSE_LOWERING: "pauseM1M3RaisingLowering",
+                self.TEXT_RESUME_LOWERING: "resumeM1M3RaisingLowering",
                 self.TEXT_ENTER_ENGINEERING: "enterEngineering",
                 self.TEXT_EXIT_ENGINEERING: "exitEngineering",
             },
             "detailedState",
             "detailedState",
         )
+        self.m1m3 = m1m3
+
+        self._panic_button = ColoredButton(self.TEXT_PANIC)
+        self._panic_button.setColor(Qt.red)
+        self._panic_button.setSizePolicy(QSizePolicy.Preferred, QSizePolicy.Expanding)
+        self._panic_button.setMinimumHeight(50)
+        self._panic_button.clicked.connect(self._panic)
+
+        self.insert_widget(self._panic_button, 0)
+
+        self.m1m3.detailedState.connect(self.detailed_state)
 
     def get_state_buttons_map(self, state: int) -> list[str | None]:
         states_map: dict[int, list[str | None]] = {
             DetailedStates.STANDBY: [
                 self.TEXT_START,
+                None,
                 None,
                 None,
                 None,
@@ -278,14 +300,16 @@ class M1M3CSCControl(CSCControlWidget):
                 self.TEXT_ENABLE,
                 None,
                 None,
+                None,
                 self.TEXT_STANDBY,
             ],
-            DetailedStates.FAULT: [self.TEXT_STANDBY, None, None, None, None],
-            DetailedStates.OFFLINE: [None, None, None, None, None],
+            DetailedStates.FAULT: [self.TEXT_STANDBY, None, None, None, None, None],
+            DetailedStates.OFFLINE: [None, None, None, None, None, None],
             DetailedStates.PARKED: [
                 None,
                 self.TEXT_DISABLE,
                 self.TEXT_RAISE,
+                None,
                 self.TEXT_ENTER_ENGINEERING,
                 None,
             ],
@@ -293,6 +317,7 @@ class M1M3CSCControl(CSCControlWidget):
                 None,
                 self.TEXT_DISABLE,
                 self.TEXT_RAISE,
+                None,
                 self.TEXT_EXIT_ENGINEERING,
                 None,
             ],
@@ -300,6 +325,7 @@ class M1M3CSCControl(CSCControlWidget):
                 None,
                 self.TEXT_ABORT_RAISE,
                 None,
+                self.TEXT_PAUSE_RAISING,
                 None,
                 None,
             ],
@@ -307,6 +333,7 @@ class M1M3CSCControl(CSCControlWidget):
                 None,
                 self.TEXT_ABORT_RAISE,
                 None,
+                self.TEXT_PAUSE_RAISING,
                 None,
                 None,
             ],
@@ -314,6 +341,7 @@ class M1M3CSCControl(CSCControlWidget):
                 None,
                 None,
                 self.TEXT_LOWER,
+                None,
                 self.TEXT_ENTER_ENGINEERING,
                 None,
             ],
@@ -321,15 +349,63 @@ class M1M3CSCControl(CSCControlWidget):
                 None,
                 None,
                 self.TEXT_LOWER,
+                None,
                 self.TEXT_EXIT_ENGINEERING,
                 None,
             ],
-            DetailedStates.LOWERING: [None, None, None, None, None],
-            DetailedStates.LOWERINGENGINEERING: [None, None, None, None, None],
-            DetailedStates.LOWERINGFAULT: [None, None, None, None, None],
+            DetailedStates.LOWERING: [
+                None,
+                None,
+                None,
+                self.TEXT_PAUSE_LOWERING,
+                None,
+                None,
+            ],
+            DetailedStates.LOWERINGENGINEERING: [
+                None,
+                None,
+                None,
+                self.TEXT_PAUSE_LOWERING,
+                None,
+                None,
+            ],
+            DetailedStates.LOWERINGFAULT: [None, None, None, None, None, None],
             DetailedStates.PROFILEHARDPOINTCORRECTIONS: [
                 None,
                 None,
+                None,
+                None,
+                None,
+                None,
+            ],
+            DetailedStates.PAUSEDRAISING: [
+                None,
+                None,
+                self.TEXT_RESUME_RAISING,
+                None,
+                None,
+                None,
+            ],
+            DetailedStates.PAUSEDRAISINGENGINEERING: [
+                None,
+                None,
+                self.TEXT_RESUME_RAISING,
+                None,
+                None,
+                None,
+            ],
+            DetailedStates.PAUSEDLOWERING: [
+                None,
+                None,
+                self.TEXT_RESUME_LOWERING,
+                None,
+                None,
+                None,
+            ],
+            DetailedStates.PAUSEDLOWERINGENGINEERING: [
+                None,
+                None,
+                self.TEXT_RESUME_LOWERING,
                 None,
                 None,
                 None,
@@ -337,6 +413,16 @@ class M1M3CSCControl(CSCControlWidget):
         }
 
         return states_map[state]
+
+    @asyncSlot()
+    async def _panic(self, checked: bool = False) -> None:
+        await command(self, self.m1m3.remote.cmd_panic)
+
+    @Slot()
+    def detailed_state(self, data: typing.Any) -> None:
+        self._panic_button.setEnabled(
+            not (data.detailedState == DetailedStates.OFFLINE)
+        )
 
 
 class ApplicationControlWidget(QWidget):
@@ -346,8 +432,6 @@ class ApplicationControlWidget(QWidget):
     changes.
     """
 
-    TEXT_PANIC = "&Panic!"
-
     def __init__(self, m1m3: MetaSAL):
         super().__init__()
 
@@ -356,12 +440,6 @@ class ApplicationControlWidget(QWidget):
 
         command_layout = QVBoxLayout()
 
-        self._panic_button = ColoredButton(self.TEXT_PANIC)
-        self._panic_button.setColor(Qt.red)
-        self._panic_button.setSizePolicy(QSizePolicy.Preferred, QSizePolicy.Expanding)
-        self._panic_button.clicked.connect(self._panic)
-
-        command_layout.addWidget(self._panic_button)
         command_layout.addWidget(M1M3CSCControl(m1m3))
 
         self.supportedNumber = QLCDNumber(6)
@@ -384,8 +462,6 @@ class ApplicationControlWidget(QWidget):
         self.slewWidget.setEnabled(False)
         command_layout.addWidget(self.slewWidget)
 
-        command_layout.addStretch()
-
         self.weightSupportedPercent = QProgressBar()
         self.weightSupportedPercent.setOrientation(Qt.Vertical)
         self.weightSupportedPercent.setRange(0, 100)
@@ -404,15 +480,8 @@ class ApplicationControlWidget(QWidget):
         self.m1m3.hardpointMonitorData.connect(self.hardpointMonitorData)
         self.m1m3.hardpointActuatorSettings.connect(self.hardpointActuatorSettings)
 
-    @asyncSlot()
-    async def _panic(self, checked: bool = False) -> None:
-        await command(self, self.m1m3.remote.cmd_panic)
-
     @Slot()
     def detailedState(self, data: typing.Any) -> None:
-        self._panic_button.setEnabled(
-            not (data.detailedState == DetailedStates.OFFLINE)
-        )
         self.slewWidget.setEnabled(
             not (
                 data.detailedState